Walls and score are now shown
The board is now lined with walls. The aim arrow has a new texture and the
placement of shot and next shot are updated.

The panel now has its own background and there is also a chalk board
texture for the score board where the current score is shown. Points are
given for shooting balls and slightly more for dangling balls.