Shots only have a valid colour, win is possible
It is now possible to generate a valid colour, that is, a colour that is
in use on the board. This is used when generating colours for the shots.
New rows added at the top can always have all the colours specified for
the game (in Ball.Initialize()).

Winning is possible when all balls have been shot. End screen has been
added but is empty.