Highscore screen is fully functional
There is now a screen showing all the highscores for all the levels and
difficulties. Buttons on both menu and end screen to show it.

HighscoreList was taken out of EndScreen into its own class since it is
used in two screens (EndScreen and HighscoreScreen).

Tweaked ball colours.

DVDCover added, manual and manual cover updated.