Options screen added
A new options screen has been added along with a new button on the start
screen to access it. The screen has checkboxes for the different settings
and the settings are now global in the Core class. There is also
information about shortcut keys.

Checkbox class used on the options screen.

Highscore screen now has information text and a sign when the focus board
is shown, hiding the buttons. The main highscore screen has information
text both on the chalkboards and below them.

There is a longer aim "shadow" added and options to toggle its showing.